Add files via upload
Updated the evaluation function - with dynamic weighting. So far I am factoring position and material.
However based on the amount of material lost, the weighting will be different. For similar material, the position will be weighted more. For vast material difference, the material difference will be factored more. I have used a sigmoid function to get the evaluation for material difference, since 1 point difference is quite minute, however 3 points is losing. Anything above 9-39 points is going to be extremely losing.
I fixed the issues of having only one arraylist in the entire program, to allocating new ones for future moves. I brushed up on some other stuff and now the entire program works with no runtime errors! However there are some bugs because some moves aren't being reverted back in the correct color. There are some issues for when the colors are flipped.